initial refactor to remove ermrest_config module
