+++ deriva/core/__init__.py
@@ -5,7 +5,7 @@ from deriva.core.base_cli import BaseCLI, KeyValuePairArgs
 from deriva.core.deriva_binding import DerivaBinding, DerivaPathError, DerivaClientContext
 from deriva.core.deriva_server import DerivaServer
 from deriva.core.ermrest_catalog import ErmrestCatalog, ErmrestSnapshot, ErmrestCatalogMutationError
-from deriva.core.ermrest_config import AttrDict, CatalogConfig
+from deriva.core.ermrest_model import AttrDict, Model, Schema, Table, Column, Key, ForeignKey
 from deriva.core.polling_ermrest_catalog import PollingErmrestCatalog
 from deriva.core.hatrac_store import HatracStore, HatracHashMismatch, HatracJobPaused, HatracJobAborted, \
     HatracJobTimeout
