+++ docs/README.md
@@ -342,17 +342,14 @@ factory methods on a catalog or snapshot object:
 - `path_builder = catalog.getPathBuilder()`
    - `deriva.core.datapath.Catalog` path builder for catalog (or snapshot)
    - Allows higher-level data access idioms as described previously.
-- `config_root = catalog.getCatalogConfig()`
-   - `deriva.core.ermrest_config.CatalogConfig` object for catalog (or snapshot)
-   - The `config_root` object roots a tree of objects isomorphic to the catalog model, organizing configuration data according to each part of the model.
-   - Allows inspection of catalog/snapshot annotations and policies.
-   - Allows mutation to draft a new configuration objective.
-   - Draft changes are applied with `config_root.apply()`
 - `model_root = catalog.getCatalogModel()`
    - `deriva.core.ermrest_model.Model` object for catalog (or snapshot)
    - The `model_root` object roots a tree of objects isomorphic to the catalog model, organizing model definitions according to each part of the model.
    - Allows inspection of catalog/snapshot models (schemas, tables, columns, constraints)
-   - Some model management idioms are exposed as methods on individual objects in the model hierarchy.
+   - Allows inspection of catalog/snapshot annotations and policies.
+   - Allows configuration field mutation to draft a new configuration objective.
+   - Draft changes are applied with `model_root.apply()`
+   - Many model management idioms are exposed as methods on individual objects in the model hierarchy.
 
 ### Low-Level HTTP Methods
